organise file structure and store token in local storage
